pyproject.toml
[project]
name = "flaskapp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Soul", email = ""},
]
dependencies = ["Flask>=3.0.3", "websocket-client>=1.8.0"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
